requirements.txt
# Federated Learning-based Adaptive Traffic Signal Control System
# Python Dependencies

# Federated Learning Framework
flwr>=1.25.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.8.0

# Data Processing
numpy>=2.0.0
pandas>=2.0.0

# Visualization
matplotlib>=3.8.0

# Utilities
pyyaml>=6.0.0
